Prevent terraform apply step to be canceled.
If the step "HashiCorp - Terraform Apply" is canceled during the
execution, this can prevent the infra's state from being saved, and as
such the resources created cannot be destroyed correctly.  To prevent
unhandled resources (such as droplets, DNS records, etc.), the step is
protected from being canceled by adding the `success()` and the
`always()` expressions to the `if` statement.  The step is executed only
if all the previous steps are succeded, by the success() expression, and
if it is running, cannot be canceled thanks to the expression
`always()`.